+++ src/app/service/constants/channelIds.ts
@@ -1,9 +1,5 @@
-export default Object.freeze({
-	bountyBoard: process.env.DISCORD_CHANNEL_BOUNTY_BOARD_ID,
-	scoapSquad: process.env.DISCORD_CHANNEL_SCOAP_SQUAD_ID,
-	COMMUNITY_CALLS_STAGE: process.env.DISCORD_STAGE_COMMUNITY_CALL_ID,
-	DEV_WORKROOM: process.env.DISCORD_CHANNEL_DEV_WORKROOM_ID,
-	WRITERS_ROOM: process.env.DISCORD_CHANNEL_WRITERS_ROOM_ID,
-	generalSupport: process.env.DISCORD_CHANNEL_SUPPORT_ID,
-	firstQuestProject: process.env.DISCORD_CHANNEL_FIRST_QUEST_PROJECT_ID,
+const channelIds = Object.freeze({
+	DM: 'DIRECT_MESSAGE',
 });
+
+export default channelIds;