+++ src/app/utils/tracer.ts
@@ -1,3 +1,0 @@
-import tracer from 'dd-trace';
-tracer.init();
-export default tracer;